+++ src/OpenTelemetry.Exporter.Prometheus/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* `UsePrometheus` extension method for adding middleware to an ASP.NET Core
+   application moved from `OpenTelemetry.Exporter.Prometheus` namespace to
+   `Microsoft.AspNetCore.Builder` namespace.
+  ([#1576](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1576))
+
 ## 1.0.0-rc1.1
 
 Released 2020-Nov-17